requirements.txt
telethon>=1.34.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
